chore: move element scripts
