--- Runtime/AvatarCreator/Elements.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: a0d9288cdb8da8c4b844f9c14d9385f7
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
