+++ Runtime/AvatarCreator/UI/ImageConfirmationElement.cs.meta
