+++ Runtime/AvatarCreator/UI/PhotoCaptureElement.cs.meta
